+++ src/main/java/roomescape/reservation/presentation/controller/web/ReservationController.java
@@ -1,0 +1,13 @@
+package roomescape.reservation.presentation.controller.web;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class ReservationController {
+
+    @GetMapping("/admin/reservation")
+    public String reservation() {
+        return "admin/reservation";
+    }
+}
